refactor: SEO 친화적 URL 패턴 적용(#314)
