+++ src/features/chatting-page/components/ChatRoomInfo.tsx
@@ -49,7 +49,7 @@ export function ChatRoomInfo({ data, onLeaveRoom, onBack }: ChatRoomInfoProps) {
         </Button>
       </div>
       <Link
-        href={ROUTES.DETAIL_ID(Number(data?.productId))}
+        href={ROUTES.DETAIL_ID(Number(data?.productId), data?.productTitle)}
         className="flex items-center gap-2 rounded-lg border border-gray-200 px-2.5 py-3 hover:bg-gray-50"
       >
         <ChatProductCard productImageUrl={data?.productImageUrl} productTitle={data?.productTitle} productPrice={data?.productPrice} size="md" />
